refactor: address daily quality findings for 2026-08-12
Extract Create Workspace setup subscription into a hook and pure helpers;
split Settings multi-route screens out of a single hub file; archive
quality score 80 report.
