+++ apps/mobile/src/features/workspaces/CreateWorkspaceScreen.tsx
@@ -1,6 +1,6 @@
 import { Button, Host } from "@expo/ui";
 import { expoUiButtonStretchModifiers } from "@/ui/primitives/expo-ui-button-modifiers";
-import { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { useEffect, useMemo, useState } from "react";
 import { Platform, StyleSheet, Text, View } from "react-native";
 import { useRouter } from "expo-router";
 import { useMutation, useQuery } from "@tanstack/react-query";
@@ -11,11 +11,15 @@ import {
   getCreateWorkspaceReadiness,
   selectCreateWorkspaceProjectGuid,
 } from "@/features/workspaces/create-workspace-readiness";
+import {
+  formatSetupStatus,
+  slugify,
+} from "@/features/workspaces/create-workspace-helpers";
+import { useCreateWorkspaceSetup } from "@/features/workspaces/use-create-workspace-setup";
 import { useMobileWs } from "@/providers/MobileWsProvider";
 
 import { useSessionStore } from "@/stores/session-store";
-import { isWorkspaceSetupProgressNotification, wsActions } from "@/api/ws-actions";
-import type { WorkspaceModel, WorkspaceSetupProgressNotification, WsNotification } from "@/api/types";
+import { wsActions } from "@/api/ws-actions";
 import { colors, radii } from "@/theme/colors";
 import { useMobileTheme } from "@/theme/theme-store";
 import {
@@ -45,12 +49,6 @@ const STATUS_OPTIONS = [
 ];
 
 const EMPTY_LABEL_VALUE = "__none__";
-const ANSI_PATTERN = /\x1B\[[0-?]*[ -/]*[@-~]/g;
-
-type SetupSnapshot = {
-  progress: WorkspaceSetupProgressNotification;
-  output: string;
-};
 
 export function CreateWorkspaceScreen({ initialProjectGuid }: { initialProjectGuid?: string | null }) {
   const router = useRouter();
@@ -72,12 +70,14 @@ export function CreateWorkspaceScreen({ initialProjectGuid }: { initialProjectGu
   const [selectedLabelGuids, setSelectedLabelGuids] = useState<string[]>([]);
   const [labelToAdd, setLabelToAdd] = useState(EMPTY_LABEL_VALUE);
   const [error, setError] = useState<string | null>(null);
-  const [createdWorkspace, setCreatedWorkspace] = useState<WorkspaceModel | null>(null);
-  const [isAwaitingSetup, setIsAwaitingSetup] = useState(false);
-  const [setupProgress, setSetupProgress] = useState<WorkspaceSetupProgressNotification | null>(null);
-  const [setupOutput, setSetupOutput] = useState("");
-  const createdWorkspaceIdRef = useRef<string | null>(null);
-  const setupSnapshotsRef = useRef(new Map<string, SetupSnapshot>());
+  const {
+    beginAwaitingSetup,
+    createdWorkspace,
+    isAwaitingSetup,
+    setIsAwaitingSetup,
+    setupOutput,
+    setupProgress,
+  } = useCreateWorkspaceSetup({ client, setError });
 
   const bootstrap = useQuery({
     queryKey: ["workspace-bootstrap", selectedServerId, state],
@@ -115,63 +115,6 @@ export function CreateWorkspaceScreen({ initialProjectGuid }: { initialProjectGu
       setProjectGuid(nextProjectGuid);
     }
   }, [initialProjectGuid, projectGuid, projectOptions]);
-
-  const applySetupSnapshot = useCallback(
-    (snapshot: SetupSnapshot) => {
-      const { progress, output } = snapshot;
-      setSetupProgress(progress);
-      setSetupOutput(output);
-
-      if (progress.status === "completed" && progress.success) {
-        setError(null);
-        setIsAwaitingSetup(false);
-        router.replace(`/workspace/${progress.workspace_id}`);
-        return;
-      }
-
-      if (progress.status === "error" || !progress.success) {
-        setIsAwaitingSetup(false);
-        setError(progress.output ? cleanSetupOutput(progress.output).trim() : "Workspace setup failed.");
-        return;
-      }
-
-      setError(null);
-      setIsAwaitingSetup(true);
-    },
-    [router],
-  );
-
-  const recordSetupProgress = useCallback(
-    (progress: WorkspaceSetupProgressNotification) => {
-      const previous = setupSnapshotsRef.current.get(progress.workspace_id);
-      const incomingOutput = progress.output ? cleanSetupOutput(progress.output) : "";
-      const output = progress.output
-        ? progress.replace_output
-          ? incomingOutput
-          : `${previous?.output ?? ""}${incomingOutput}`
-        : previous?.output ?? "";
-      const snapshot = { progress, output };
-
-      setupSnapshotsRef.current.set(progress.workspace_id, snapshot);
-
-      if (createdWorkspaceIdRef.current === progress.workspace_id) {
-        applySetupSnapshot(snapshot);
-      }
-    },
-    [applySetupSnapshot],
-  );
-
-  useEffect(() => {
-    if (!client) return;
-    const unsubscribe = client.subscribeMessages((message) => {
-      if (!isWsNotification(message) || message.payload.event !== "workspace_setup_progress") return;
-      if (!isWorkspaceSetupProgressNotification(message.payload.data)) return;
-      recordSetupProgress(message.payload.data);
-    });
-    return () => {
-      unsubscribe();
-    };
-  }, [client, recordSetupProgress]);
 
   const createWorkspace = useMutation({
     mutationFn: async () => {
@@ -211,31 +154,7 @@ export function CreateWorkspaceScreen({ initialProjectGuid }: { initialProjectGu
       return workspace;
     },
     onSuccess: (workspace) => {
-      createdWorkspaceIdRef.current = workspace.guid;
-      setCreatedWorkspace(workspace);
-      setError(null);
-
-      const cachedSnapshot = setupSnapshotsRef.current.get(workspace.guid);
-      if (cachedSnapshot) {
-        applySetupSnapshot(cachedSnapshot);
-        return;
-      }
-
-      setSetupOutput("");
-      setSetupProgress({
-        workspace_id: workspace.guid,
-        status: "creating",
-        step_key: "create_worktree",
-        failed_step_key: null,
-        step_title: "Preparing Workspace",
-        output: null,
-        replace_output: false,
-        requires_confirmation: false,
-        success: true,
-        countdown: null,
-        setup_context: null,
-      });
-      setIsAwaitingSetup(true);
+      beginAwaitingSetup(workspace);
     },
     onError: (nextError) => {
       setError(nextError instanceof Error ? nextError.message : "Workspace creation failed.");
@@ -587,37 +506,6 @@ export function CreateWorkspaceScreen({ initialProjectGuid }: { initialProjectGu
 
       <InlineError message={error ?? (bootstrap.error instanceof Error ? bootstrap.error.message : null)} />
     </AppScreen>
-  );
-}
-
-function isWsNotification(message: unknown): message is WsNotification {
-  if (!message || typeof message !== "object") return false;
-  const envelope = message as Record<string, unknown>;
-  if (envelope.type !== "notification") return false;
-  const payload = envelope.payload as Record<string, unknown> | undefined;
-  return Boolean(payload && typeof payload.event === "string" && "data" in payload);
-}
-
-function cleanSetupOutput(value: string) {
-  return value.replace(ANSI_PATTERN, "").replace(/\r/g, "");
-}
-
-function formatSetupStatus(progress: WorkspaceSetupProgressNotification) {
-  if (progress.requires_confirmation) return "Waiting for confirmation";
-  if (progress.status === "completed") return "Completed";
-  if (progress.status === "error") return "Failed";
-  if (progress.status === "setting_up") return "Running setup";
-  return "Creating";
-}
-
-function slugify(value: string) {
-  return (
-    value
-      .trim()
-      .toLowerCase()
-      .replace(/[^a-z0-9]+/g, "-")
-      .replace(/^-+|-+$/g, "")
-      .slice(0, 48) || `mobile-${Date.now()}`
   );
 }
